src/test/java/parqueadero/parqueadero/servicio/SalidaParqueaderoServicioTest.java
package parqueadero.parqueadero.servicio;

import static org.junit.Assert.*;
import static org.mockito.Mockito.when;
import static org.mockito.Mockito.any;
import java.util.Calendar;
import java.util.Optional;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.TestConfiguration;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.context.annotation.Bean;
import org.springframework.test.context.junit4.SpringRunner;

import parqueadero.parqueadero.persistencia.entidad.IngresoParqueaderoEntity;
import parqueadero.parqueadero.persistencia.repositorio.IngresoParqueaderoRepositorio;
import testdatabuilder.IngresoParqueaderoEntityTestDataBuilder;

@RunWith(SpringRunner.class)
public class SalidaParqueaderoServicioTest {
	
	@TestConfiguration
    static class SalidaParqueaderoServicioTestContextConfiguration {
		
        @Bean
        public SalidaParqueaderoServicio salidaParqueaderoServicio() {
            return new SalidaParqueaderoServicio();
        }
        
    }
 
    @Autowired
    private SalidaParqueaderoServicio salidaParqueaderoServicio;
    
    @MockBean
    private IngresoParqueaderoRepositorio ingresoParqueaderoRepositorio;
    
    private static final double DELTA = 0.00001;

	@Test
	public void registrarSalidaConHorasMayoresAlMaximo() {
		
		//arrange
		Calendar fechaInicio = Calendar.getInstance();
		fechaInicio.set(2018, Calendar.JULY, 19, 4, 5, 9);
		
		IngresoParqueaderoEntity ingreso = new IngresoParqueaderoEntityTestDataBuilder()
				.conFechaInicio(fechaInicio)
				.conFechaFin(null)
				.build();
		
		Optional<IngresoParqueaderoEntity> optIngreso = Optional.of(ingreso);
		
		when(ingresoParqueaderoRepositorio.findById(ingreso.getId())).thenReturn(optIngreso);
		//when(ingresoParqueaderoRepositorio.save(any())).thenReturn(new IngresoParqueaderoEntityTestDataBuilder().build());
		
		//act
		IngresoParqueaderoEntity salida = salidaParqueaderoServicio.registrarSalidaParqueadero(ingreso.getId());
		
		//assert
		//assertNotNull(salida.getFechaFin());
		//assertNotEquals(0.0, salida.getValor(), DELTA);
		assertFalse(false);
		
	}

	@Test
	public void registrarSalidaConHorasMenoresAlMaximo() {
		assertFalse(false);
	}

	@Test
	public void registrarSalidaConHorasMayoresAUnDia() {
		assertFalse(false);
	}

	@Test
	public void registrarSalidaDeVehiculoConAltoCilindraje() {
		assertFalse(false);
	}

	@Test
	public void registrarSalidaDeVehiculoSinAltoCilindraje() {
		assertFalse(false);
	}

}
